+++ checker/src/main/java/org/checkerframework/checker/nullness/NullnessNoInitSubchecker.java
@@ -20,6 +20,18 @@ import java.util.Set;
  *
  * <p>The {@link NullnessChecker} uses this checker as the target (see {@link
  * InitializationChecker#getTargetCheckerClass()}) for its initialization type system.
+ *
+ * <p>You can use the following {@link SuppressWarnings} prefixes with this checker:
+ *
+ * <ul>
+ *   <li>{@code @SuppressWarnings("nullness")} suppresses warnings for nullness, initialization and
+ *       keyfor annotations
+ *   <li>{@code @SuppressWarnings("nullnessnokeyfor")} suppresses warnings for nullness and
+ *       initialization annotations only
+ *   <li>{@code @SuppressWarnings("nullnessnoinit")} suppresses warnings for nullness and keyfor
+ *       annotations only
+ *   <li>{@code @SuppressWarnings("nullnessonly")} suppresses warnings for nullness annotations only
+ * </ul>
  */
 @StubFiles({"junit-assertions.astub"})
 public class NullnessNoInitSubchecker extends BaseTypeChecker {