+++ checker/src/main/java/org/checkerframework/checker/nullness/NullnessChecker.java
@@ -21,7 +21,7 @@ import javax.annotation.processing.SupportedOptions;
  * <p>You can use the following {@link SuppressWarnings} prefixes with this checker:
  *
  * <ul>
- *   <li>{@code @SuppressWarnings("nullness")} suppresses warnings for nullness, initialization and
+ *   <li>{@code @SuppressWarnings("nullness")} suppresses warnings for nullness, initialization, and
  *       keyfor annotations
  *   <li>{@code @SuppressWarnings("nullnessnokeyfor")} suppresses warnings for nullness and
  *       initialization annotations only