+++ docs/CHANGELOG.md
@@ -17,13 +17,17 @@ The Checker Framework runs under JDK 25 -- that is, it runs on a version 25 JVM.
 
 **Closed issues:**
 
-eisop#1051, eisop#1115, eisop#1180.
+eisop#1051, eisop#1059, eisop#1115, eisop#1180.
 
 
 Version 3.49.1-eisop1 (March 17, 2025)
 --------------------------------------
 
 **User-visible changes:**
+
+All checkers issue `annotation.on.supertype` error for annotation write on
+supertype in class declaration. Other support annotations on supertype can override
+`BaseTypeVisitor#checkSupertypeAnnotations(Tree)` to disable the check.
 
 The Nullness Checker now reports an error if any instanceof pattern variables
 are annotated with `@Nullable` and a redundant warning if they are annotated