+++ docs/CHANGELOG.md
@@ -13,7 +13,7 @@ select a different JDK for the actual compilation and testing.
 
 **Closed issues:**
 
-eisop#1051, eisop#1115, eisop#1180.
+eisop#1051, eisop#1059, eisop#1115, eisop#1180.
 
 
 Version 3.49.3 (May 2, 2025)
@@ -41,6 +41,10 @@ Version 3.49.1-eisop1 (March 17, 2025)
 --------------------------------------
 
 **User-visible changes:**
+
+All checkers issue `annotation.on.supertype` error for annotation write on
+supertype in class declaration. Other support annotations on supertype can override
+`BaseTypeVisitor#checkSupertypeAnnotations(Tree)` to disable the check.
 
 The Nullness Checker now reports an error if any instanceof pattern variables
 are annotated with `@Nullable` and a redundant warning if they are annotated